+++ AENetwork.xcodeproj/AENetwork.h
@@ -1,0 +1,10 @@
+/**
+ *  https://github.com/tadija/AENetwork
+ *  Copyright (c) Marko Tadić 2017
+ *  Licensed under the MIT license. See LICENSE file.
+ */
+
+#import <Foundation/Foundation.h>
+
+FOUNDATION_EXPORT double AENetworkVersionNumber;
+FOUNDATION_EXPORT const unsigned char AENetworkVersionString[];