+++ Tests/LinuxMain.swift
@@ -1,6 +1,12 @@
+/**
+ *  https://github.com/tadija/AENetwork
+ *  Copyright (c) Marko Tadić 2017
+ *  Licensed under the MIT license. See LICENSE file.
+ */
+
 import XCTest
-@testable import NetworkTests
+@testable import AENetworkTests
 
 XCTMain([
-     testCase(NetworkTests.allTests),
+     testCase(AENetworkTests.allTests),
 ])